safari-600.7-branch/LayoutTests/fast/forms/menulist-no-renderer-for-unexpected-children-expected.txt

Check that a select control does not render children that are not <option> or <optgroup>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.innerText is ""
PASS successfullyParsed is true

TEST COMPLETE
